Added blockchain serialize

+++ include/transaction.h
@@ -30,7 +30,7 @@ typedef struct transaction_t {
     time_t created_at;
     ssh_key_t sender_public_key;
     ssh_key_t recipient_public_key;
-    uint32_t amount;
+    uint64_t amount;
     ssh_signature_t sender_signature;
 } transaction_t;
 
@@ -51,7 +51,7 @@ return_code_t transaction_create(
     transaction_t **transaction,
     ssh_key_t *sender_public_key,
     ssh_key_t *recipient_public_key,
-    uint32_t amount,
+    uint64_t amount,
     ssh_key_t *sender_private_key
 );
